+++ src/pages/Dashboard/components/CardDetail/CardDetailStyled.ts
